Moving table view cells to view folder
