+++ HexaCalc/View/SelectionSummaryTableViewCell.swift
