+++ HexaCalc/View/SelectionTableViewCell.swift
