+++ HexaCalc/View/TextTableViewCell.swift
